requirement.txt
mcp>=1.0.0
google-generativeai>=0.8.0
Pillow>=10.0.0
python-dotenv